feature issue #13726 change to showTextCollapsed parameter for Menu.Item and Menu.SubMenu
