+++ components/menu/SubMenu.tsx
@@ -20,12 +20,21 @@ class SubMenu extends React.Component<any, any> {
     this.subMenu = subMenu;
   };
 
+  getCollapsedMode = () => {
+    const { mode, showTextCollapsed } = this.props;
+    return mode === 'vertical' && showTextCollapsed === true;
+  };
+
   render() {
     const { rootPrefixCls, className } = this.props;
     const theme = this.context.antdMenuTheme;
+    const subMenuClassName = classNames(className, {
+      [`${rootPrefixCls}-submenu-collapse`]: this.getCollapsedMode(),
+    });
     return (
       <RcSubMenu
         {...this.props}
+        className={subMenuClassName}
         ref={this.saveSubMenu}
         popupClassName={classNames(`${rootPrefixCls}-${theme}`, className)}
       />
